Solve Travis CI failures
Travis CI needs extra permissions on the Docker container for systemd to
work properly, it seems
inspec gem now seems to not include the binary which has been split into
its own package "inspec-bin". So we add that to the run.
Re-order builds so that the longest running one starts ealiest. This
improves the amount of parallel capacity we can exercise within the
limits of the Travis environment
Switch to /usr/sbin/init instead of calling systemd directly